simplify earnings calculator
