+++ apps/web/lib/zod/schemas/program-lander.ts
@@ -63,8 +63,6 @@ export const programLanderEarningsCalculatorBlockSchema =
   programLanderBlockCommonSchema.extend({
     type: z.literal("earnings-calculator"),
     data: z.object({
-      title: programLanderBlockTitleSchema,
-      description: programLanderBlockDescriptionSchema,
       productPrice: z.number(),
     }),
   });
